Move sidebar to a component.
