+++ src/components/Sidebar/Sidebar.css
@@ -1,0 +1,10 @@
+.sidebar {
+  box-sizing: border-box;
+  height: 100%;
+  width: 30vw;
+  padding: 0 1rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  overflow: hidden;
+}
