--- src/components/Version/util.ts
@@ -1,5 +1,0 @@
-import type { Metadata } from "./Version";
-
-export function getLatestVersion(metadata: Metadata) {
-  return metadata.latest ?? "latest";
-}
